+++ packages/timeline-state-resolver/src/integrations/casparCG/index.ts
@@ -1,7 +1,7 @@
 import * as _ from 'underscore'
 import * as deepMerge from 'deepmerge'
 import { DeviceWithState, CommandWithContext, DeviceStatus, StatusCode, literal } from '../../devices/device'
-import { CasparCG, Command as CommandNS, AMCPUtil, AMCP, CasparCGSocketStatusEvent } from 'casparcg-connection'
+import { AMCPCommand, BasicCasparCGAPI, Commands, Response } from 'casparcg-connection'
 import {
 	DeviceType,
 	TimelineContentTypeCasparCg,
@@ -10,9 +10,7 @@ import {
 	TimelineContentCCGProducerBase,
 	ResolvedTimelineObjectInstanceExtended,
 	DeviceOptionsCasparCG,
-	Transition,
 	Mappings,
-	TSRTransitionOptions,
 	TimelineContentCasparCGAny,
 	TSRTimelineObjProps,
 	Timeline,
@@ -21,8 +19,7 @@ import {
 
 import {
 	CasparCGState,
-	AMCPCommandVOWithContext,
-	ChannelInfo,
+	AMCPCommandWithContext,
 	LayerBase,
 	MediaLayer,
 	InputLayer,
@@ -38,6 +35,7 @@ import {
 	Transition as StateTransition,
 	Mixer,
 } from 'casparcg-state'
+import { InternalState } from 'casparcg-state/dist/lib/stateObjectStorage'
 import { DoOnTime, SendMode } from '../../devices/doOnTime'
 import * as request from 'request'
 import { InternalTransitionHandler } from '../../devices/transitions/transitionHandler'
@@ -45,19 +43,12 @@ import Debug from 'debug'
 import { endTrace, startTrace } from '../../lib'
 const debug = Debug('timeline-state-resolver:casparcg')
 
-const MAX_TIMESYNC_TRIES = 5
-const MAX_TIMESYNC_DURATION = 40
 const MEDIA_RETRY_INTERVAL = 10 * 1000 // default time in ms between checking whether a file needs to be retried loading
 
 export interface DeviceOptionsCasparCGInternal extends DeviceOptionsCasparCG {
 	commandReceiver?: CommandReceiver
 }
-export type CommandReceiver = (
-	time: number,
-	cmd: CommandNS.IAMCPCommand,
-	context: string,
-	timelineObjId: string
-) => Promise<any>
+export type CommandReceiver = (time: number, cmd: AMCPCommand, context: string, timelineObjId: string) => Promise<any>
 /**
  * This class is used to interface with CasparCG installations. It creates
  * device states from timeline states and then diffs these states to generate
@@ -65,13 +56,8 @@ export type CommandReceiver = (
  * optionally, uses the CasparCG command scheduling features.
  */
 export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCGInternal> {
-	private _ccg: CasparCG
-	private _ccgState: CasparCGState
-	private _queue: { [token: string]: { time: number; command: CommandNS.IAMCPCommand } } = {}
+	private _ccg: BasicCasparCGAPI
 	private _commandReceiver: CommandReceiver
-	private _timeToTimecodeMap: { time: number; timecode: number } = { time: 0, timecode: 0 }
-	private _timeBase: { [channel: string]: number } | number = {}
-	private _useScheduling?: boolean
 	private _doOnTime: DoOnTime
 	private initOptions?: CasparCGOptions
 	private _connected = false
@@ -79,6 +65,7 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 	private _transitionHandler: InternalTransitionHandler = new InternalTransitionHandler()
 	private _retryTimeout: NodeJS.Timeout
 	private _retryTime: number | null = null
+	private _currentState: InternalState = { channels: {} }
 
 	constructor(deviceId: string, deviceOptions: DeviceOptionsCasparCGInternal, getCurrentTime: () => Promise<number>) {
 		super(deviceId, deviceOptions, getCurrentTime)
@@ -86,11 +73,8 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 		if (deviceOptions.options) {
 			if (deviceOptions.commandReceiver) this._commandReceiver = deviceOptions.commandReceiver
 			else this._commandReceiver = this._defaultCommandReceiver.bind(this)
-
-			if (deviceOptions.options.timeBase) this._timeBase = deviceOptions.options.timeBase
 		}
 
-		this._ccgState = new CasparCGState()
 		this._doOnTime = new DoOnTime(
 			() => {
 				return this.getCurrentTime()
@@ -107,42 +91,52 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 	 */
 	async init(initOptions: CasparCGOptions): Promise<boolean> {
 		this.initOptions = initOptions
-		this._useScheduling = initOptions.useScheduling
-		this._ccg = new CasparCG({
+		this._ccg = new BasicCasparCGAPI({
 			host: initOptions.host,
 			port: initOptions.port,
-			autoConnect: true,
-			virginServerCheck: true,
-			onConnectionChanged: (connected: boolean) => {
-				this._connected = connected
-				this._connectionChanged()
-			},
 		})
 
-		this._ccg.on(CasparCGSocketStatusEvent.CONNECTED, (event: CasparCGSocketStatusEvent) => {
+		this._ccg.on('connect', () => {
 			this.makeReady(false) // always make sure timecode is correct, setting it can never do bad
 				.catch((e) => this.emit('error', 'casparCG.makeReady', e))
-			if (event.valueOf().virginServer === true) {
-				// a "virgin server" was just restarted (so it is cleared & black).
-				// Otherwise it was probably just a loss of connection
 
-				this._ccgState.softClearState()
-				this.clearStates()
-				this.emit('resetResolver')
-			}
+			this._connected = true
+			this._connectionChanged()
+
+			// TODO - maybe add this back based on info command
+			// if (event.valueOf().virginServer === true) {
+			// 	// a "virgin server" was just restarted (so it is cleared & black).
+			// 	// Otherwise it was probably just a loss of connection
+
+			// 	this._currentState = { channels: {} }
+			// 	this.clearStates()
+			// 	this.emit('resetResolver')
+			// }
 		})
 
-		const command = await this._ccg.info()
-		this._ccgState.initStateFromChannelInfo(
-			_.map(command.response.data, (obj: any) => {
-				return {
+		this._ccg.on('disconnect', () => {
+			this._connected = false
+			this._connectionChanged()
+		})
+
+		const { error, request } = await this._ccg.executeCommand({ command: Commands.Info, params: {} })
+		if (error) {
+			return false // todo - should this throw?
+		}
+		const response = await request
+
+		if (response?.data[0]) {
+			response.data.forEach((obj) => {
+				this._currentState.channels[obj.channel] = {
 					channelNo: obj.channel,
 					videoMode: obj.format.toUpperCase(),
 					fps: obj.frameRate,
+					layers: {},
 				}
-			}) as ChannelInfo[],
-			this.getCurrentTime()
-		)
+			})
+		} else {
+			return false // not being able to get channel count is a problem for us
+		}
 
 		if (typeof initOptions.retryInterval === 'number' && initOptions.retryInterval >= 0) {
 			this._retryTime = initOptions.retryInterval || MEDIA_RETRY_INTERVAL
@@ -165,32 +159,22 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 				return
 			}
 			this._ccg.disconnect()
-			this._ccg.onDisconnected = () => {
+			this._ccg.once('disconnect', () => {
 				resolve(true)
-			}
+			})
 		})
 	}
 	/** Called by the Conductor a bit before a .handleState is called */
 	prepareForHandleState(newStateTime: number) {
 		// Clear any queued commands later than this time:
-		if (this._useScheduling) {
-			// Can't do it
-			// this._clearScheduledFutureCommands(newStateTime, commandsToAchieveState)
-		} else {
-			this._doOnTime.clearQueueNowAndAfter(newStateTime)
-			this.cleanUpStates(0, newStateTime)
-		}
+		this._doOnTime.clearQueueNowAndAfter(newStateTime)
+		this.cleanUpStates(0, newStateTime)
 	}
 	/**
 	 * Generates an array of CasparCG commands by comparing the newState against the oldState, or the current device state.
 	 */
 	handleState(newState: Timeline.TimelineState<TSRTimelineContent>, newMappings: Mappings) {
 		super.onHandleState(newState, newMappings)
-		// check if initialized:
-		if (!this._ccgState.isInitialised) {
-			this.emit('warning', 'CasparCG State not initialized yet')
-			return
-		}
 
 		const previousStateTime = Math.max(this.getCurrentTime(), newState.time)
 
@@ -201,15 +185,16 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 		this.emit('timeTrace', endTrace(convertTrace))
 
 		const diffTrace = startTrace(`device:diffState`, { deviceId: this.deviceId })
-		const commandsToAchieveState = this._diffStates(oldCasparState, newCasparState, newState.time)
+		const commandsToAchieveState = CasparCGState.diffStatesOrderedCommands(
+			oldCasparState as InternalState,
+			newCasparState,
+			newState.time
+		)
 		this.emit('timeTrace', endTrace(diffTrace))
 
 		// clear any queued commands later than this time:
-		if (this._useScheduling) {
-			this._clearScheduledFutureCommands(newState.time, commandsToAchieveState)
-		} else {
-			this._doOnTime.clearQueueNowAndAfter(previousStateTime)
-		}
+		this._doOnTime.clearQueueNowAndAfter(previousStateTime)
+
 		// add the new commands to the queue:
 		this._addToQueue(commandsToAchieveState, newState.time)
 
@@ -222,17 +207,7 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 	 * @param clearAfterTime
 	 */
 	clearFuture(clearAfterTime: number) {
-		if (this._useScheduling) {
-			for (const token in this._queue) {
-				if (this._queue[token].time > clearAfterTime) {
-					this._doCommand(new AMCP.ScheduleRemoveCommand(token), `clearFuture (${clearAfterTime})`, '').catch((e) =>
-						this.emit('error', 'CasparCG.ScheduleRemoveCommand', e)
-					)
-				}
-			}
-		} else {
-			this._doOnTime.clearQueueAfter(clearAfterTime)
-		}
+		this._doOnTime.clearQueueAfter(clearAfterTime)
 	}
 	get canConnect(): boolean {
 		return true
@@ -250,14 +225,6 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 			return 'CasparCG ' + this.deviceId + ' ' + this._ccg.host + ':' + this._ccg.port
 		} else {
 			return 'Uninitialized CasparCG ' + this.deviceId
-		}
-	}
-
-	get queue() {
-		if (this._queue) {
-			return _.map(this._queue, (val, index) => [val, index])
-		} else {
-			return []
 		}
 	}
 
@@ -325,11 +292,10 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 					channelLayout: content.channelLayout,
 					format: content.deviceFormat,
 				},
-				filter: content.filter,
 				playing: true,
 				playTime: null,
 
-				vfilter: content.videoFilter,
+				vfilter: content.videoFilter || content.filter,
 				afilter: content.audioFilter,
 			})
 		} else if (content.type === TimelineContentTypeCasparCg.TEMPLATE) {
@@ -426,7 +392,8 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 					if (baseContent.transitions.outTransition) {
 						transitions.outTransition = new StateTransition(baseContent.transitions.outTransition)
 					}
-					stateLayer.media = new TransitionObject(media, {
+					// todo - not a fan of this type assertion but think it's ok
+					stateLayer.media = new TransitionObject(media as string, {
 						inTransition: transitions.inTransition,
 						outTransition: transitions.outTransition,
 					})
@@ -583,55 +550,23 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 	 */
 	async makeReady(okToDestroyStuff?: boolean): Promise<void> {
 		// Sync Caspar Time to our time:
-		const command = await this._ccg.info()
-		const channels: any[] = command.response.data
-		const attemptSync = async (channelNo, tries): Promise<void> => {
-			const startTime = this.getCurrentTime()
-			await this._commandReceiver(
-				startTime,
-				new AMCP.TimeCommand({
-					channel: channelNo,
-					timecode: this.convertTimeToTimecode(startTime, channelNo),
-				}),
-				'makeReady',
-				''
-			)
+		const command = await this._ccg.executeCommand({ command: Commands.Info, params: {} })
+		if (command.error) throw new Error('Could not makeReady')
+		const response = await command.request
+		const channels: any[] = response.data
 
-			const duration = this.getCurrentTime() - startTime
-			if (duration > MAX_TIMESYNC_DURATION) {
-				// @todo: acceptable time is dependent on fps
-				if (tries > MAX_TIMESYNC_TRIES) {
-					this.emit(
-						'error',
-						'CasparCG',
-						new Error(
-							`CasparCG Time command took too long (${MAX_TIMESYNC_TRIES} tries took longer than ${MAX_TIMESYNC_DURATION}ms), channel will be slightly out of sync!`
-						)
-					)
-					return Promise.resolve()
-				}
-				await new Promise<void>((resolve) => {
-					setTimeout(() => resolve(), MAX_TIMESYNC_DURATION)
-				})
-				await attemptSync(channelNo, tries + 1)
-			}
-		}
-
-		if (this._useScheduling) {
-			for (const channel of channels) {
-				const channelNo = channel.channel
-				await attemptSync(channelNo, 1)
-			}
-		}
 		// Clear all channels (?)
 		if (okToDestroyStuff) {
 			await Promise.all(
 				_.map(channels, async (channel: any) => {
 					await this._commandReceiver(
 						this.getCurrentTime(),
-						new AMCP.ClearCommand({
-							channel: channel.channel,
-						}),
+						{
+							command: Commands.Clear,
+							params: {
+								channel: channel.channel,
+							},
+						},
 						'makeReady and destroystuff',
 						''
 					)
@@ -681,11 +616,6 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 			}
 		}
 
-		if (!this._ccgState.isInitialised) {
-			statusCode = StatusCode.BAD
-			messages.push(`CasparCG device connection not initialized (restart required)`)
-		}
-
 		if (this._queueOverflow) {
 			statusCode = StatusCode.BAD
 			messages.push('Command queue overflow: CasparCG server has to be restarted')
@@ -698,114 +628,21 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 		}
 	}
 	/**
-	 * Compares the new timeline-state with the old one, and generates commands to account for the difference
-	 */
-	private _diffStates(oldState, newState, time: number): Array<AMCPCommandVOWithContext> {
-		// @todo: this is a tmp fix for the command order. should be removed when ccg-state has been refactored.
-		return CasparCGState.diffStatesOrderedCommands(oldState, newState, time)
-	}
-	private async _doCommand(command: CommandNS.IAMCPCommand, context: string, timlineObjId: string): Promise<void> {
-		const time = this.getCurrentTime()
-
-		const interceptedCommand = this._interceptCommand(command)
-
-		if (interceptedCommand) {
-			return this._commandReceiver(time, interceptedCommand, context, timlineObjId)
-		} else {
-			return Promise.resolve()
-		}
-	}
-	/**
-	 * Clear future commands after {@code time} if they are not in {@code commandsToSendNow}.
-	 */
-	private _clearScheduledFutureCommands(time: number, commandsToSendNow: Array<AMCPCommandVOWithContext>) {
-		// clear any queued commands later than this time:
-		const now = this.getCurrentTime()
-
-		_.each(this._queue, (q, token: string) => {
-			if (q.time < now) {
-				// the command has expired / been executed
-				delete this._queue[token]
-			} else if (q.time >= time) {
-				// The command is in the future
-
-				// check if that command is about to be scheduled here as well:
-				let matchingCommand: CommandNS.IAMCPCommand | undefined
-				let matchingCommandI = -1
-				if (q.time === time) {
-					_.each(commandsToSendNow, (cmd: CommandNS.IAMCPCommandVO, i) => {
-						const command: CommandNS.IAMCPCommand = AMCPUtil.deSerialize(cmd, 'id')
-
-						if (
-							command.name === q.command.name &&
-							command.channel === q.command.channel &&
-							command.layer === q.command.layer &&
-							_.isEqual(command.payload, q.command.payload)
-						) {
-							matchingCommand = command
-							matchingCommandI = i
-						}
-					})
-				}
-
-				if (matchingCommand) {
-					// We're about to send a command that's already scheduled in CasparCG
-					// just ignore it then..
-
-					// remove the commands from commands to send
-					commandsToSendNow.splice(matchingCommandI, 1)
-				} else {
-					this._doCommand(new AMCP.ScheduleRemoveCommand(token), `_clearScheduledFutureCommands (${time})`, '').catch(
-						(e) => this.emit('error', 'CasparCG.ScheduleRemoveCommand', e)
-					)
-					delete this._queue[token]
-				}
-			}
-		})
-	}
-	/**
 	 * Use either AMCP Command Scheduling or the doOnTime to execute commands at
 	 * {@code time}.
 	 * @param commandsToAchieveState Commands to be added to queue
 	 * @param time Point in time to send commands at
 	 */
-	private _addToQueue(commandsToAchieveState: Array<AMCPCommandVOWithContext>, time: number) {
-		let i = 0
-		const now = this.getCurrentTime()
-
-		_.each(commandsToAchieveState, (cmd: AMCPCommandVOWithContext) => {
-			const command = AMCPUtil.deSerialize(cmd, 'id')
-
-			if (this._useScheduling) {
-				if (time <= now) {
-					this._doCommand(command, cmd.context.context, cmd.context.layerId).catch((e) =>
-						this.emit('error', 'CasparCG._doCommand', e)
-					)
-				} else {
-					const token = `${time.toString(36).substr(-8)}_${('000' + i++).substr(-4)}`
-					const scheduleCommand = new AMCP.ScheduleSetCommand({
-						token,
-						timecode: this.convertTimeToTimecode(time, command.channel),
-						command,
-					})
-					this._doCommand(scheduleCommand, cmd.context.context, cmd.context.layerId).catch((e) =>
-						this.emit('error', 'CasparCG._doCommand', e)
-					)
-					this._queue[token] = {
-						time: time,
-						command: command,
-					}
-				}
-			} else {
-				this._doOnTime.queue(
-					time,
-					undefined,
-					async (c: { command: CommandNS.IAMCPCommand; cmd: AMCPCommandVOWithContext }) => {
-						return this._doCommand(c.command, c.cmd.context.context, c.cmd.context.layerId)
-					},
-					{ command: command, cmd: cmd }
-				)
-			}
+	private _addToQueue(commandsToAchieveState: Array<AMCPCommandWithContext>, time: number) {
+		_.each(commandsToAchieveState, (cmd: AMCPCommandWithContext) => {
+			this._doOnTime.queue(
+				time,
+				undefined,
+				async (c: { command: AMCPCommand; cmd: AMCPCommandWithContext }) => {
+					return this._commandReceiver(time, c.command, c.cmd.context.context, c.cmd.context.layerId)
+				},
+				{ command: { command: cmd.command, params: cmd.params }, cmd: cmd }
+			)
 		})
 	}
 	/**
@@ -815,7 +652,7 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 	 */
 	private async _defaultCommandReceiver(
 		time: number,
-		cmd: CommandNS.IAMCPCommand,
+		cmd: AMCPCommand,
 		context: string,
 		timelineObjId: string
 	): Promise<any> {
@@ -832,83 +669,113 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 		}
 		this.emitDebug(cwc)
 
-		return this._ccg
-			.do(cmd)
-			.then((resCommand) => {
-				if (this._queue[resCommand.token]) {
-					delete this._queue[resCommand.token]
+		const { request, error } = await this._ccg.executeCommand(cmd)
+		if (error) {
+			this.emit('commandError', error, cwc)
+		}
+
+		try {
+			const response = await request
+
+			// I forgot what this means.. oh well... todo
+			if (!response) return
+
+			this._changeTrackedStateFromCommand(cmd, response, time)
+
+			if (response.responseCode === 504 && !this._queueOverflow) {
+				this._queueOverflow = true
+				this._connectionChanged()
+			} else if (this._queueOverflow) {
+				this._queueOverflow = false
+				this._connectionChanged()
+			}
+
+			if (response.responseCode >= 400) {
+				// this is an error code:
+				let errorString = `${response.responseCode} ${cmd.command} ${response.type}: ${response.type}`
+
+				if (Object.keys(cmd.params).length) {
+					errorString += ' ' + JSON.stringify(cmd.params)
 				}
-				// If the command was performed successfully, copy the state from the current state into the tracked caspar-state:
-				// This is later used in _assertIntendedState
-				if (
-					(resCommand.name === 'LoadbgCommand' ||
-						resCommand.name === 'PlayCommand' ||
-						resCommand.name === 'LoadCommand' ||
-						resCommand.name === 'ClearCommand' ||
-						resCommand.name === 'StopCommand' ||
-						resCommand.name === 'ResumeCommand') &&
-					resCommand.channel &&
-					resCommand.layer
-				) {
-					const currentState = this.getState(time)
-					if (currentState) {
-						const currentCasparState = currentState.state
 
-						const trackedState = this._ccgState.getState()
+				this.emit('commandError', new Error(errorString), cwc)
+			}
+		} catch (e) {
+			// This shouldn't really happen
+			this.emit('commandError', Error('Command not sent: ' + e), cwc)
+		}
+	}
 
-						const channel = currentCasparState.channels[resCommand.channel]
-						if (channel) {
-							if (!trackedState.channels[resCommand.channel]) {
-								trackedState.channels[resCommand.channel] = {
-									channelNo: channel.channelNo,
-									fps: channel.fps || 0,
-									videoMode: channel.videoMode || null,
-									layers: {},
-								}
-							}
-							// Copy the tracked from current state:
-							trackedState.channels[resCommand.channel].layers[resCommand.layer] = channel.layers[resCommand.layer]
-							this._ccgState.setState(trackedState)
+	private _changeTrackedStateFromCommand(command: AMCPCommand, response: Response, time: number) {
+		if (
+			response.responseCode < 300 && // TODO - maybe we accept every code except 404?
+			response.command.match(/Loadbg|Play|Load|Clear|Stop|Resume/i) &&
+			'channel' in command.params &&
+			command.params.channel !== undefined &&
+			'layer' in command.params &&
+			command.params.layer !== undefined
+		) {
+			const currentState = this.getState(time)
+			if (currentState) {
+				const currentCasparState = currentState.state
+
+				const trackedState = this._currentState
+
+				const channel = currentCasparState.channels[command.params.channel]
+				if (channel) {
+					if (!trackedState.channels[command.params.channel]) {
+						trackedState.channels[command.params.channel] = {
+							channelNo: channel.channelNo,
+							fps: channel.fps || 0,
+							videoMode: channel.videoMode || null,
+							layers: {},
+						}
+					}
+
+					// copy into the trackedState
+					if (
+						(command.command === Commands.Play && command.params.clip) ||
+						(!('clip' in command.params) &&
+							trackedState.channels[command.params.channel].layers[command.params.layer].nextUp)
+					) {
+						// a play command without parameters (channel/layer) is only succesful if the nextUp worked
+						// a play command with params can always be accepted
+						trackedState.channels[command.params.channel].layers[command.params.layer] = {
+							...channel.layers[command.params.layer],
+							nextUp: undefined, // a play command always clears nextUp
+						}
+					} else if (command.command === Commands.Loadbg) {
+						// only loadbg can set nextUp and nextUp can only be set by loadbg
+						trackedState.channels[command.params.channel].layers[command.params.layer] = {
+							...trackedState.channels[command.params.channel].layers[command.params.layer],
+							nextUp: channel.layers[command.params.layer].nextUp,
+						}
+					} else if (
+						command.command === Commands.Stop &&
+						trackedState.channels[command.params.channel].layers[command.params.layer].nextUp?.auto
+					) {
+						// auto next + stop means bg -> fg => nextUp cleared
+						trackedState.channels[command.params.channel].layers[command.params.layer] = {
+							...channel.layers[command.params.layer],
+							nextUp: undefined, // auto next + stop means bg -> fg => nextUp cleared
+						}
+					} else if (command.command === Commands.Resume || command.command === Commands.Stop) {
+						// stop and resume can be done without affecting nextup
+						trackedState.channels[command.params.channel].layers[command.params.layer] = {
+							...channel.layers[command.params.layer],
+							nextUp: trackedState.channels[command.params.channel].layers[command.params.layer].nextUp,
+						}
+					} else {
+						// anything else can always be copied but also clears nextUp
+						// @todo - can LOADBG be followed by an empty LOAD? (if yes, apply same logic as PLAY)
+						trackedState.channels[command.params.channel].layers[command.params.layer] = {
+							...channel.layers[command.params.layer],
+							nextUp: undefined,
 						}
 					}
 				}
-
-				if (this._queueOverflow) {
-					this._queueOverflow = false
-					this._connectionChanged()
-				}
-			})
-			.catch((error) => {
-				if (error?.response?.code === 504) {
-					if (!this._queueOverflow) {
-						this._queueOverflow = true
-						this._connectionChanged()
-					}
-				}
-
-				let errorString = ''
-				if (error?.response?.code === 404) {
-					errorString = `404: File not found`
-				}
-
-				if (!errorString) {
-					errorString = error?.response?.raw ? error.response.raw : error.toString()
-				}
-
-				if (cmd.name) {
-					errorString += ` ${cmd.name} `
-				}
-				if (cmd['_objectParams'] && !_.isEmpty(cmd['_objectParams'])) {
-					errorString += ', params: ' + JSON.stringify(cmd['_objectParams'])
-				} else if (cmd.payload && !_.isEmpty(cmd.payload)) {
-					errorString += ', payload: ' + JSON.stringify(cmd.payload)
-				}
-				this.emit('commandError', new Error(errorString), cwc)
-				if (cmd.name === 'ScheduleSetCommand') {
-					// delete this._queue[cmd.getParam('command').token]
-					delete this._queue[cmd.token]
-				}
-			})
+			}
+		}
 	}
 
 	/**
@@ -927,16 +794,17 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 
 		const ccgState = tlState.state
 
-		const diff = this._ccgState.getDiff(ccgState, this.getCurrentTime())
+		const diff = CasparCGState.diffStates(this._currentState, ccgState, this.getCurrentTime())
 
-		const cmd: Array<AMCPCommandVOWithContext> = []
+		const cmd: Array<AMCPCommandWithContext> = []
 		for (const layer of diff) {
 			// filter out media commands
 			for (let i = 0; i < layer.cmds.length; i++) {
 				if (
-					layer.cmds[i]._commandName === 'LoadbgCommand' ||
-					(layer.cmds[i]._commandName === 'PlayCommand' && layer.cmds[i]._objectParams.clip) ||
-					layer.cmds[i]._commandName === 'LoadCommand'
+					// todo - shall we pass decklinks etc. as well?
+					layer.cmds[i].command === Commands.Loadbg ||
+					layer.cmds[i].command === Commands.Load ||
+					(layer.cmds[i].command === Commands.Play && 'clip' in (layer.cmds[i].params as Record<string, unknown>))
 				) {
 					layer.cmds[i].context.context += ' [RETRY]'
 					cmd.push(layer.cmds[i])
@@ -949,226 +817,7 @@ export class CasparCGDevice extends DeviceWithState<State, DeviceOptionsCasparCG
 		}
 	}
 
-	/**
-	 * Converts ms to timecode.
-	 * @param time Time to convert
-	 * @param channel Channel to use for timebase
-	 */
-	private convertTimeToTimecode(time: number, channel: number): string {
-		const relTime = time - this._timeToTimecodeMap.time
-		const timecodeTime = this._timeToTimecodeMap.timecode + relTime
-
-		const timeBase =
-			(typeof this._timeBase === 'object' ? this._timeBase[channel + ''] : this._timeBase) ||
-			(this.initOptions ? this.initOptions.fps || 25 : 25)
-
-		const timecode = [
-			('0' + (Math.floor(timecodeTime / 3.6e6) % 24)).substr(-2),
-			('0' + (Math.floor(timecodeTime / 6e4) % 60)).substr(-2),
-			('0' + (Math.floor(timecodeTime / 1e3) % 60)).substr(-2),
-			('0' + (Math.floor(timecodeTime / (1000 / timeBase)) % timeBase)).substr(-(timeBase + '').length),
-		]
-
-		return timecode.join(':')
-	}
 	private _connectionChanged() {
 		this.emit('connectionChanged', this.getStatus())
-	}
-	/**
-	 * Intercept the casparcg-connection commands, for internal transitions
-	 * Returns the command if it's not intercepted along the way
-	 */
-	_interceptCommand(command: CommandNS.IAMCPCommand): CommandNS.IAMCPCommand | undefined {
-		// Intercept internal commands:
-		const objectParams: any = command['_objectParams']
-		if (objectParams) {
-			const transitionOptions: TSRTransitionOptions = objectParams.customOptions as TSRTransitionOptions
-
-			if (transitionOptions) {
-				if (objectParams.keyword === 'FILL') {
-					if (objectParams.transition === Transition.TSR_TRANSITION) {
-						// Handle transitions internally:
-
-						this._transitionHandler.activateTransition(
-							this._getTransitionId('FILL', command.channel, command.layer),
-							[0, 0, 1, 1],
-							[objectParams.x, objectParams.y, objectParams.xScale, objectParams.yScale],
-							['position', 'position', 'scale', 'scale'],
-							transitionOptions,
-							{
-								position: {
-									type: 'physical',
-								},
-								scale: {
-									type: 'linear',
-									options: {
-										linearSpeed: 0.0003, // tmp: todo: remove hard-coding of this
-									},
-								},
-							},
-							(newValues) => {
-								const c = new AMCP.MixerFillCommand({
-									channel: command.channel,
-									layer: command.layer,
-									x: newValues[0],
-									y: newValues[1],
-									xScale: newValues[2],
-									yScale: newValues[3],
-								})
-								this._commandReceiver(this.getCurrentTime(), c, 'Internal transition', 'internalTransition').catch(
-									(e) => this.emit('error', 'CasparCG.InternalTransition', e)
-								)
-							}
-						)
-						// Abort: don't send the original command
-						return undefined
-					} else {
-						this._transitionHandler.stopAndSnapTransition(
-							this._getTransitionId('FILL', command.channel, command.layer),
-							[objectParams.x, objectParams.y, objectParams.xScale, objectParams.yScale]
-						)
-					}
-				} else if (objectParams.keyword === 'PERSPECTIVE') {
-					if (objectParams.transition === Transition.TSR_TRANSITION) {
-						// Handle transitions internally:
-
-						this._transitionHandler.activateTransition(
-							this._getTransitionId('PERSPECTIVE', command.channel, command.layer),
-							[0, 0, 1, 0, 1, 1, 0, 1],
-							[
-								objectParams.topLeftX,
-								objectParams.topLeftY,
-								objectParams.topRightX,
-								objectParams.topRightY,
-								objectParams.bottomRightX,
-								objectParams.bottomRightY,
-								objectParams.bottomLeftX,
-								objectParams.bottomLeftY,
-							],
-							['tl', 'tl', 'tr', 'tr', 'br', 'br', 'bl', 'bl'],
-							transitionOptions,
-							{
-								tl: { type: 'physical' }, // top left corner
-								tr: { type: 'physical' }, // top right corner
-								bl: { type: 'physical' }, // bottom right corner
-								br: { type: 'physical' }, // bottom left corner
-							},
-							(newValues) => {
-								const c = new AMCP.MixerPerspectiveCommand({
-									channel: command.channel,
-									layer: command.layer,
-									topLeftX: newValues[0],
-									topLeftY: newValues[1],
-									topRightX: newValues[2],
-									topRightY: newValues[3],
-									bottomRightX: newValues[4],
-									bottomRightY: newValues[5],
-									bottomLeftX: newValues[6],
-									bottomLeftY: newValues[7],
-								})
-								this._commandReceiver(this.getCurrentTime(), c, 'Internal transition', 'internalTransition').catch(
-									(e) => this.emit('error', 'CasparCG.InternalTransition', e)
-								)
-							}
-						)
-						// Abort: don't send the original command
-						return undefined
-					} else {
-						this._transitionHandler.stopAndSnapTransition(
-							this._getTransitionId('PERSPECTIVE', command.channel, command.layer),
-							[
-								objectParams.topLeftX,
-								objectParams.topLeftY,
-								objectParams.topRightX,
-								objectParams.topRightY,
-								objectParams.bottomRightX,
-								objectParams.bottomRightY,
-								objectParams.bottomLeftX,
-								objectParams.bottomLeftY,
-							]
-						)
-					}
-				} else if (objectParams.keyword === 'OPACITY' || objectParams.keyword === 'VOLUME') {
-					const opt: { initial: number; prop: string } =
-						objectParams.keyword === 'OPACITY'
-							? {
-									initial: 1,
-									prop: 'opacity',
-							  }
-							: objectParams.keyword === 'VOLUME'
-							? {
-									initial: 1,
-									prop: 'volume',
-							  }
-							: {
-									initial: 0,
-									prop: 'N/A',
-							  }
-
-					if (objectParams.transition === Transition.TSR_TRANSITION) {
-						// Handle transitions internally:
-
-						this._transitionHandler.activateTransition(
-							this._getTransitionId(objectParams.keyword, command.channel, command.layer),
-							[opt.initial],
-							[objectParams[opt.prop]],
-							['v'],
-							transitionOptions,
-							{
-								v: { type: 'linear' }, // tmp hack: for these, a linear would be better that physical
-							},
-							(newValues) => {
-								const properties = {
-									channel: command.channel,
-									layer: command.layer,
-								}
-								properties[opt.prop] = newValues[0]
-								const c = new AMCP[command.name](properties)
-								this._commandReceiver(this.getCurrentTime(), c, 'Internal transition', 'internalTransition').catch(
-									(e) => this.emit('error', 'CasparCG.InternalTransition', e)
-								)
-							}
-						)
-						// Abort: don't send the original command
-						return undefined
-					} else {
-						this._transitionHandler.stopAndSnapTransition(
-							this._getTransitionId(objectParams.keyword, command.channel, command.layer),
-							[objectParams[opt.prop]]
-						)
-					}
-				} else if (objectParams.keyword === 'CLEAR') {
-					if (command.layer) {
-						this._getTransitions(undefined, command.channel, command.layer).forEach((identifier) =>
-							this._transitionHandler.clearTransition(identifier)
-						)
-					} else {
-						// Clear the whole channel:
-						this._getTransitions(undefined, command.channel).forEach((identifier) =>
-							this._transitionHandler.clearTransition(identifier)
-						)
-					}
-				}
-			}
-		}
-		return command
-	}
-	private _getTransitionId(keyword: string, channel: number, layer?: number) {
-		return `${keyword}_${channel}-${layer || ''}`
-	}
-	private _getTransitions(keyword?: string, channel?: number, layer?: number): string[] {
-		let regex = ''
-		if (keyword) {
-			regex = `^${keyword}_`
-		}
-		if (channel) {
-			regex += `_${channel}-`
-		}
-		if (layer) {
-			regex += `-${layer}$`
-		}
-		regex = regex.replace(/__/, '_').replace(/--/, '--')
-
-		return this._transitionHandler.getIdentifiers().filter((i) => i.match(new RegExp(regex)))
 	}
 }