+++ CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.5.0-release47.3](https://github.com/nrkno/sofie-timeline-state-resolver/compare/7.5.0-release47.2...7.5.0-release47.3) (2022-11-07)
+
+**Note:** Version bump only for package timeline-state-resolver-packages
+
+
+
+
+
+# [7.5.0-release47.2](https://github.com/nrkno/tv-automation-state-timeline-resolver/compare/7.5.0-release47.1...7.5.0-release47.2) (2022-11-02)
+
+**Note:** Version bump only for package timeline-state-resolver-packages
+
+
+
+
+
+# [7.5.0-release47.1](https://github.com/nrkno/tv-automation-state-timeline-resolver/compare/7.5.0-release47.0...7.5.0-release47.1) (2022-11-02)
+
+
+### Bug Fixes
+
+* add method to manually purge viz rundown ([49737e2](https://github.com/nrkno/tv-automation-state-timeline-resolver/commit/49737e2fba1b967b1b5d6d84e5c1624ee3a9ab11))
+
+
+
+
+
 # [7.5.0-release47.0](https://github.com/nrkno/sofie-timeline-state-resolver/compare/7.3.0...7.5.0-release47.0) (2022-10-28)